Backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# SQLAlchemy & Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Python utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
